Fix issue with invites to MUC if a MUC PM for that room is open
Previously if you wanted to invite people to a MUC and had
a PM window for a MUC occupant open at the same time, the
InviteToMUCUIEvent would be handled by the PM window, by the
ChatController of the PM window and not the MUCController of
the MUC window.

Test-Information:

Verified that some scenarios work correctly:
- Tested a drop to a MUC window while a MUC PM window is open
  to an occupant in the MUC. Previously this crashed due to
  ChatsManager::localMUCServiceJID_ being empty.

- Test that impromptu MUC creation to a normal chat works.
- Test that impromptu MUC creation to a MUC PM chat works.

All unit and integration tests pass on macOS 10.12.1.

Change-Id: Ib20de7e925e3503308211936ee47d4ba829d0394