Do not emit windowClosing signal in QtTabbable dtor
Instead, handle the QCloseEvent and emit the windowClosing
signal there. This is an attempt to fix a crash I am unable
to reproduce but which crash stack includes the signal emit
in ~QtTabbable.

Test-Information:

Builds on macOS 10.12.1 with clang master. All unit and
integration tests pass. Manual login and close of Swift worked
without a crash.

Change-Id: I470f3c5362fd0f502c542560854424e87fc0727e