Sluift: Add client methods for adding and removing JIDs from blocklist
A previous commit added a get_block_list() client method to retrieve
the list of blocked items.
This patch extends that and adds capabilities to add and remove
items to such list.

Test-information:

Tested with client and was able to successfully add items to list,
delete and delete all
Also tried to delete a non existing item and it rightly complained.

Change-Id: Iabbfdbd9e74ddc9740db5c9871b8355210aa0727