Add alternative zoom shortcuts on macOS and workaround Qt bug
This adds the Ctrl + Minus and Ctrl + Equal shortcuts for
zoom out and zoom in respectively.

In addition, this commit provides a workaround for QTBUG-56571.

Test-Information:

Tested on macOS 10.12.3 with Qt 5.7.1. Tested with German
and UK keyboard layout. On UK layout both shortcuts,
Ctrl + (+/-) and Ctrl + (=/-) work. On German layout only
Ctrl + (+/-) works.

Change-Id: Ifbcab308c9a8f4c88b51978969c73c3c1138a9ba