Improve Swift about window regarding auto update UX
The dialog will automatically initiate a check for updates
when opened.

It will show the current state of the auto updater backend,
e.g. whether it is downloading or already at the latest
version.

This also fixes update channel selection being shown
on Windows and Linux.

Test-Information:

Ran Swift and opening the dialog shows a short progress bar
indicating it checking for updates. Afterwards it shows that
it is already running the latest version, which is sensible
for a dev build.

Change-Id: Ie08cd2a8852e468d6007122604b532fedc24bcfe