Use FlowLayout instead of QGridLayout in QtEmojiGrid
FlowLayout is an official BSD-licensed Qt example showing
how to implement custom layouts. It will layout items
dynamically in rows.

This way we don’t need static column/row calculations for
QGridLayout and it looks better.

Test-Information:

Build and ran on macOS 10.12.3 with Qt 5.7 to test that it has
a better, less spacious look.

Change-Id: Ief1299b0d3fb1e516a1973469f4f9a26824942f2