Add Sparkle license to about dialog COPYING file on demand
If scons detects a Sparkle release in the 3rdParty folder,
it will add the containing LICENSE file at the end of the
dynamically generated COPYING file.

Test-Information:

Tested that the contents of Sparkle’s LICENSE file are at
the end of the text in the license window of the about dialog,
if the Sparkle folder, with a Sparkle release inside, is
present in 3rdParty.

Tested that Swift builds fine, if the Sparkle folder is
missing in 3rdParty.

Change-Id: I89db155b5bb3eef4d632997fd99ff270b4720a16