Allow pasting a newline separated list of JIDs in blocking dialog
Furthermore the dialog will alphabetically sort the list
of blocked JIDs before showing it in the dialog.

Test-Information:

Tested on macOS 10.12.3 with Qt 5.7.1 that it sorts the
list it receives from the server before presenting it in the
UI. Successfully pasted a newline spereated list of JIDs in
the dialog.

Change-Id: I8fe6969821bccad0193180cee9433da43285aaef