Do no emit onXMPPDataRead when the session is terminated
On a bosh error BOSHSessionStream, posts an event to the event
loop that closes the XMPP stream. If onXMPPDataRead is
emitted, that data is handled before the close of the stream
though, which can cause another BOSH request to be sent.

Test-Information:

Added a unit test verifying the behaviour.

Wrote a custom python script replicating the original error
and verified that Swift does not stay connected on a BOSH
error with additional text data.

Tested on macOS 10.12.3 with recent clang trunk.

Change-Id: Ie90099afa0934707a6758b00706a65227ceb48b8