Fix incorrect default chat view font size
A chat font size value of 2 represents the native default
font size on the system, which should be the default font
size for the chat view on first install. We allow this to be
reduced to 0 and increased without a limit.

Furthermore the settings key was changed so that Swift 3 users
will automatically use the default size when they first open
a newer Swift.

Test-Information:

Tested on Ubuntu 16.04.02 after clearing the local Swift
config.

Change-Id: I3426001eb302e22134c40bc4a479f15684ca2001