Modified XMLBeautifier to handle split payloads
When receiving network data we were processing it in chunks. Sometimes
one XML message got split across multiple chunks. XMLBeautifier was
stateless and would create a new parser for every single chunk. This was
causing multi-chunk messages to be truncated in the beautified output.
The change I made in XMLBeautifier allows it to maintain the state.

Test-Information:

Tested unger Windows 10 and CentOS.

Unit tests pass OK.

Change-Id: Idad2a8e0248ed3cbe2b47a12718b909e56ac1279