Fix for grey gap in the blue/green CSS blobs in chats
The grey gap was being created due to the relative sizes of the css
elements myavatar, myright, otheravatar, otherright.
Depending on the font size, and due to round ups, the elements ended with
a gap between them, resolving to the grey line.
The fix moves one of the elements slighly and works with newer versions of
Qt.

Test-Information:

Tested with with windows (qt5.4.2), Windows (qt5.7), Linux (qt5.5)

Change-Id: Ib4fb7fc8a5ef8bed1868a0d65707ca034309cf04