Do not grey out chat view after clearing
Test-Information:
Run Swift, open chat window, enter a message and send it. Right click on
chat view and choose Clear. Chat log in a view will be cleared but it
will stay white. On the top of the chat view will be a message: e.g.
Starting chat with test - test@test.isode.net: Offline.

Change-Id: Ie453602e2f2b14e3ecca699821521ad33937ccf6