Test against dialog type instead of nullptr in QtUserSearchWindow
If the dialog is not of type AddContact, firstMultiJIDPage_
is initialised and this states more clearly that the reset
is only done when the dialog is in this mode.

Coverity raised this issue.

Test-Information:

Added a user from search and started a chat from user search.

All unit tests pass on macOS 10.12.4.

Change-Id: I55f6df5b4f48e2c6718922846a424bf17f363b16