Don't pass the impromptu MUC 'reason' as the room JID
Changing the constructor of the CreateImpromptuMUCUIEvent class not to accept
default values to its parameters. This approach is less prone to mistakes,
like forgetting an argument.
This commit fixes SWIFT-271 bug, where only two arguments were being passed
in the constructor. As a result the reason string was used to create a room
JID, making the room unusable. By passing an empty string, the room JID
created is not valid, and when ChatsManager handles the event, it will
generate a new valid room JID.

Test information:
Unit test pass.
Tested the creation of new impromptu chat with with windows 10 (qt5.7).

Change-Id: I4ce9ee7db14b9125970a03f19d1ac0c32e1b5a0e