If silent Sparkle update impossible do not auto download updates
With this change, Swift will check at the start if it has
sufficient permissions to write to the location where
Swift is currently installed and also check if it can change
permissions of a temporary file to the permissions of the
current Swift installation.

This should prevent an authentication dialog on the exit of
Swift which does not leave a single clue to what application
the dialog belongs to.

Test-Information:

Verified that the log output indicates a possible silent update
for admin users and the impossibility of a silent update
for non-admin users and a Swift installed to /Applications
folder.

All unit tests pass on macOS 10.12.4 with Qt 5.4.2.

Change-Id: I5f03358ac67630565b3c624da157b1eeea14356d