Add update_debian_repo.sh script
The purpose of the script is as follows:

Based on an incoming folder of Debian binary and source
packages for multiple distributions and versions structured
as incoming_folder/$distribution/$version, the script will
create apt repositories using reprepro and fill it with
packages from the incoming folder structure.

If the repositories are already present, they will be updated
instead of created.

Test-Information:

Tested it on Debian 8.7 with Swift development packages
produced by our CI machine.

Change-Id: Ic0270bc954ec9609c0567d2707825fbab89e7b3a