Fix SendFile example heap-use-after-free
This fixes the crash by moving Swift::Client to a unique_ptr.
By specifying the order the objects will be destructed, we avoid
accessing the Socks5ByteStreamProxiesManager instance after it has
been destroyed. Also there is no need reset the outgoingFileTransfer
pointer.

Test Information:
Use the SendFile example to send a file to a user connected with Swift.

Change-Id: I0462dc60c1d4d5eb3b07516c6e609bbe70954039