Use QPointer to prevent access to potentially freed QtTreeWidget
As the QtTreeWidget and the QtFilterWidget are siblings in
their parent widget, it might happen that the QtTreeWidget
is deleted before the QtFilterWidget. Using Pointer, we are
able to detect this case and can prevent accessing the already
deleted QtTreeWidget.

QtFilterWidget cannot be made the child of the QtTreeWidget in
this case, so this fix uses the QPointers approach as a
workaround.

SWIFT-247

Test-Information:

Tests pass and builds on macOS 10.12.4 with Qt 5.4.2.

Change-Id: I3a60006519b580010718c4d2aa94638555c0afdf