Allow resending messages without 198 acks
This patch will allow to resend messages that have not been successfully
delivered to the server. It requires that the server will have stream
management enabled as described in XEP-198.

Test-Information:

Tested the changes on Windows and Linux. Modified the code to force
messages to fail, and then tested the resend functionality.
Added ChatControllerTest, with unit Tests for the resend case. The rest of
unit test pass.

Change-Id: Id095528247b25a47e34c5632b8469ebd4c97149b