Only apply impromptu chat title style to impromptu chats
Fixes regression of 1cb2bb7, that would show the full
bare JID as chat title for normal MUCs.

Test-Information:

Added unit test for having the classic title style for normal
MUCS.

Joined basic MUC room and verified it only shows the node
part of the MUC JID as chat title. Verified that for
impromptu chats, it shows the list of participants as
chat title.

Tested on macOS 10.12.5 with Qt 5.4.2.

Change-Id: I3685172378c0b6303524c89876b044eac1aa1422