Add scripts that test Debian packages
Adding two scripts that will test Ubuntu(xenial – yakkety)
and Debian(jessie - sid) distributions.
The script will create a base system of the distribution
and then add the swift distribution channels.
It will then test the version of the binary and source
packages that are available in the channel, and then it
will attempt to install swift and all its dependencies.

Test-Information:
Tested on Ubuntu 16.04 (host)

Change-Id: I977e72223a4688672d64b39c22c966aa143bc060