Add build_appimage.py for Linux packaging as AppImage
Run this script to generate a self-contained AppImage that
will run on any Linux distribution that has the same or newer
Linux kernel and the same or newer GLib.

The script will generate an AppImage for the same architecture
it is executed on.

By default it will use Qt5. Specify --qt5=False to create
an AppImage with Qt4.

The resulting .appimage and .debug files are copied to
./Packages/Swift

Test-Information:

Latest iteration of the script was tested with building an
AppImage on 64-bit Ubuntu 16.04 and running it on 64-bit
Debian Sid.
Earlier version of the script was tested on 32-bit SUSE Linux
Enterprise 11.

Change-Id: I14477ada3558109fe73a77f7e7fe1ee8ef629d5f