Make gtest output more compact
This patch will make the checker test application output to use a custom
printer for google test results. The output will be less verbose, and
similar to the output of CppUnit. If the --verbose flag is used, the
default printer will be used instead.

Test-information:
Tested on Windows 10 and Ubuntu 16.04.

Change-Id: I1488cf576ab07da03b0dfcc93a48a8518d5afc06