Force a build to stop if signing of the packages fails.
This patch will make the build attempt to sign the package 3 times. If
none of the attempts are successful, the build will stop and flag the
error.

Test-Information:

Tested packaging on Windows 10 (Qt 5.7), when the package signing succeeds
(signed installer package), fails (build stops) or not required (unsigned
installer package).

Change-Id: Idf1f708dda50e67e54866ca598cc5160e4ea76eb