Make the Windows installer remove older versions of Swift components
This patch will make the installer remove each file component, before
deploying the new version of the file.
This is to avoid potentials issues with the installer, when trying to
deploy a new version of a file that happens to have the same version
information as the original file. The default behaviour of the MSI
installer in a case like this is to skip the deployment of the new file.
This issue occurs when Visual Studio gets upgraded, and we need to redeploy
3rd party components of the same version but build with the new version of
the compiler.

Test-information:

Generated an installer and tested fresh install and upgrade from Swift
3.0. Tested with Windows 10 and Windows 7, normal and quiet installations.
Verified that the Qt DLLs deployed have the correct VS redist
dependencies.

Change-Id: I6f68126ee63968aad066901704a409b6ebada14f