Fix Emoji crash on insertion
Swift would crash randomly when user clicked to insert an Emoji. It was
happening because we would delete the emojisGrid_ while it was still in
use (was closing).

Test-Information:
Tested on Windows 10 and Ubuntu 16.04.
Tested by inserting an Emoji from the Emoji selector. This action does not result in
crash any more.

Change-Id: I6c0b84ea8969c7f7a4c3bfa700e345cdf63325e6