Return focus to highlight rules dialog after picking a colour
This patch fixes an issue where the Highlight and notification
configuration window was losing focus after picking a colour for some of
the highlight rules.

Test-Information:

Tested the changes and that the highlight rules dialog is no longer pushed
back after picking a colour. Tested on MacOS (10.10.5) with Qt 5.5.1 and on
Windows 10 with Qt 5.9.

Change-Id: I509bb3f79e2f8e121dafe178e52a0aac5271ee1b