Add basic TLS server support in OpenTLSContext
This also extends the TLSContext interface with methods required
for server mode.

Test-Information:

Added unit tests that test new functionality in TLSContex.
This includes test certificates in the source file that are
not for public use. This new ClientServerTest is only enabled
for OpenSSL, as other TLS backends do not support the new
functionality yet.

Tested on macOS 10.13.3 with clang-trunk.

Change-Id: I8e43476057608067eb3b9852328aa21cd22974a0