Fix a crash that occurs when inviting users in a chat
After inviting people to a chat, and converting to a MUC room, some of the
chat window signals were still signaling the original chat controller,
that has been destroying, causing a crash.

Test-Information:

Tested the changes in Windows 10 (64 bit), after triggering the
dayChangeTimer, and  onContinuationsBroken signal that was causing the
crash.

Change-Id: I70a80ab2653ed87a1dbea851157d95fb5918913f