Make carbons messages update ‘Recents’ list
This patch fixeses an issue that occurred when sending a message from
another device, the carbons messages were displayed on the chat window but
were not updating the recents chats list.

Test-Information:

Tested by logging in from two clients from the same account, and starting
a 1-1 chat with another user. Recent chats tab was updated to both
clients.

Change-Id: I6e961ebbde6a87e4bcbcc49941ab4bacbac4d9d9