Convert cppunit tests into gtests for MUCControllerTest
As stated above, and one test has been commented out as it was
not being run when it was a cppunit test and it fails due to
a bug in MUCController. The bug fix patch to follow shortly.

Test-Information:
All tests run and pass except the one mentioned above that
has been temporarily commented out.

Change-Id: Ie1ce6039c3b6c97e0072b7518b3337d5dfbf6362