Fix empty marking colour handling in MUCController
When there is a room security marking and the colours have the value ""
they should be set to the default black and white. Instead they were
being set to "", so I have updated this and uncommented the relevant unit
test.

Test-Information:
The uncommented unit test is now passing, and all other related unit
tests still pass. A MUC room with a marking but blank colour values has
been tested on a test server and works as expected.

Change-Id: I1138ca1f035f8b7c0367c36ec3a65cc857721b8c