Fix error response handling when requesting VCards
When fetching a vCard, an empty vCard response and an
item-not-found error are semantically the same. Changed the
code to treat and item-not-found error as an empty vCard in
this case.

This enables setting your own vCard on servers that do not
return an empty vCard for fresh accounts and generally
improves UX when fetching others vCards.

Test-Information:

Added unit tests verifying new behaviour. Tested with Swift
against a Prosody IM instance. Without this change you cannot
set the vCard on a fresh user. With this patch you can set your
own vCard.

Change-Id: I5f9adb4c3d6b6a1a320b834be918ab5ab0b52975