Fix convertToWindowsVersion() function to handle more RCs
The new version conversion function allows up to 8 patch releases,
up to 11 RC releases, up to 7 beta releases, and up to 5 alpha
releases.

The script fails hard on invalid input, thus failing the build
when used. It also allows for checking the release tag in the
release process.

Test-Information:

Added unit tests for some critical properties of the conversion.
As previously the minor number of the Windows version triplet
was not used, all new versions using this new scheme are
considered newer, for the same major version.

Change-Id: I53552a72ceaf7fe90f919990bd3d3be45a976c03