Mark jids with an empty domainpart as invalid
When Swiften parsed jids with an empty domainpart (e.g. user@/resource),
it would mark the jid as valid, and treat the localpart as domainpart.
RFC 6122 states in 2.2 that "A domainpart MUST NOT be zero bytes in
length".

Unit tests for both a bare jid and a full jid with an empty domainpart
have been added.

Test-Information:

Unit tests pass on Debian 9

Change-Id: Iadaf399cd4158666bfcdd6c075b8bf2102ff5538