Add timer for QtChatWindow focus to stop messages appearing unread
When a QtChatWindow's text input receives focus, it initiates a
one second timer. When it times out, the QtChatWindow checks whether
the QtChatWindow in general has focus. If so, it will reset the
unread count to 0. If it doesn't still have focus, it will not alter
the message count and the chat will still show as having unread
messages.

Test-Information:
This is in the Qt user interface, so no unit tests have been
written. Tested thoroughly on a test server on ubuntu 17.10. Tested
what happens when the chat window is changed via both a click on the
roster (both old and new) and a click on the tabs. All three scenarios
produce the desired results when both keeping focus for the full
second, or when losing focus before the second is up.

Change-Id: Idfa66990545051cfe6c9853418b2138ee0f1f57c