Add clang-tidy-fix target to Makefile
This has clang-tidy write changes out to YAML files in
parallel and finally runs clang-apply-replacements which
merges, deduplicates and applies the changes in serial.

Uses Python to determine the number of parallel jobs to run.

Test-Information:

Tested by adding new checks and running clang-tidy-fix. Works
as described above and is much faster compared to a serial
`clang-tidy -fix` run.

Change-Id: Idb357e63edeba75ef9a4fb53f5ddef2a7c3c23ee