Catch numeric cast errors in BoostConnectionServer
When BoostConnectionServer tries to set up a listener, it casts an int
to an unsigned short, which may throw an uncaught exception if the cast
fails.

Test-Information:

Unit tests pass on Debian 9
Test server code no longer crashes when it tries to set up a listener on
an invalid port.

Change-Id: If6b920e54481ce6bf174df01d14ad12eff90e3f4