Remove JID prep caching from JID class
Ran a flamegraph diff on Swift login to an account with a
moderate number of JIDs in roster and more than 10 rooms
in auto join. It showed a difference less that 0.1% on the
changed method. As it does not make much of a difference for
Swift at least, caching is removed.

Test-Information:

Unit tests passes, Swift still logins fine on macOS.

Change-Id: Id33d6d1a655580e06e1364df717fd6f34cce5327