Strip off trailing dot from domainpart of jid
RFC 6122 specifies that if a domainpart ends in a dot, it must be
stripped off before any other canonicalisation steps are taken.

Unit tests have been added to check that various JID scenarios with a
domain ending in a dot pass or are rejected as expected.

Test-Information:

Unit tests pass on macOS 10.13.
Manual tests in sluift show expected behaviour.

Change-Id: Id6813aaa4422a81bff0a4559eacd6855ef104dc3