Remove numeric_casts from XML parsers
The code has been updated to use asserts where a sensible recovery path
was deemed impossible, and a conditional return for parse.  In general,
our XML parsing will fail for any single parse of a document over
roughly 2 2GiB, which is probably not going to be a practical issue soon.

Test-Information:

Unit tests pass on macOS 10.13 using Expat and Debian 9 using LibXML

Change-Id: I3a8da802860028ea278b322af081c2b22b55a442