Address bad_numeric_casts for filetransfers
The filetransfer blockSize is now an unsigned integer, as 0 could be
used to denote an invalid block size as well (and indeed, already
indicated that better than -1 did).

All use of numeric_cast in filetransfer code has been fixed to deal with
the possibility of thrown exceptions.

Test-Information:

Unit tests pass on macOS and Debian

Change-Id: I1833d553bae071238be20ebc386ef602effb78b0