Address LinkLocal issues
Generation of TXT records might fail if any of the fields is too long,
so the result is now an optional (pending Expected). Callsites have been
updated to deal with this.

Three potentially uncaught exceptions in the Bonjour implementation have
been addressed.

Test-Information:

Unit tests pass on macOS 10.14 and Debian 9

Change-Id: Iec02c4606a18eee855362fd3c3d15614a9e72547